Fix incorrect override of Command.execute()
Take note that execute() for Attendance and Performance related commands are deleted to conform to overriding of abstract execute() method.
