+++ src/main/java/seedu/command/performance/DeletePerformance.java
@@ -10,7 +10,7 @@ import seedu.ui.Ui;
 
 import java.util.List;
 
-public class DeletePerformanceCommand extends Command {
+public class DeletePerformance extends Command {
     Performance performance;
     public String userInput;
 
@@ -19,14 +19,14 @@ public class DeletePerformanceCommand extends Command {
      * and parse it to get the Performance to be deleted.
      * @param userInput A String to be parsed.
      */
-    public DeletePerformanceCommand(String userInput) {
+    public DeletePerformance(String userInput) {
         this.userInput = userInput;
         performance = new PerformanceDataParser(userInput).getPerformance();
         PerformanceList.deletePerformance(performance);
     }
 
     @Override
-    public void execute(Ui ui, Storage storage) throws DukeException {
+    public void execute() {
 
     }
 }
