+++ src/main/java/seedu/command/performance/ViewStudentResult.java
@@ -2,6 +2,7 @@ package seedu.command.performance;
 
 import seedu.command.Command;
 import seedu.exception.DukeException;
+import seedu.exception.PerformanceStudentListException;
 import seedu.module.performance.Performance;
 import seedu.module.performance.PerformanceList;
 import seedu.storage.Storage;
@@ -9,35 +10,35 @@ import seedu.ui.Ui;
 
 import java.util.List;
 
-public class ViewAssignmentResultCommand extends Command {
+public class ViewStudentResult extends Command {
     public static List<Performance> performances;
-    private String assignment;
+    private String name;
 
     /**
      * Constructor for ViewAssignmentResultCommand. Takes String userInput
-     * and parse it to get the assignment result list to be showed.
+     * and parse it to get the student result list to be showed.
      * @param userInput A String to be parsed.
      */
-    public ViewAssignmentResultCommand(String userInput) {
+    public ViewStudentResult(String userInput) {
         String[] instructions = userInput.split(" ", 2);
-        assignment = instructions[1];
+        name = instructions[1];
         performances = new PerformanceList().getPerformanceList();
-        assignmentPerformanceList();
+        studentPerformanceList();
     }
 
     /**
      * Selects specific results of the assignment input by user
      * from the Performance list, and print the list in the format of
-     * [module][student]result.
+     * [module][assignment]result.
      */
-    public void assignmentPerformanceList() {
+    public void studentPerformanceList() {
         int size = performances.size();
         if (size == 0) {
             System.out.println("empty");
         } else {
             int i = 1;
             for (Performance performance : performances) {
-                if (performance.assignment.equals(assignment)) {
+                if (performance.nameOfStudent.equals(name)) {
                     System.out.println(i + performance.formatForStudentList());
                     i++;
                 }
@@ -46,7 +47,7 @@ public class ViewAssignmentResultCommand extends Command {
     }
 
     @Override
-    public void execute(Ui ui, Storage storage) throws DukeException {
+    public void execute() {
 
     }
 }
